+++ src/_static/setup-and-start-greenbone-community-edition.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright (C) 2022 - 2023 Greenbone AG
+# Copyright (C) 2022 - 2025 Greenbone AG
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -53,18 +53,18 @@ echo "Using Greenbone Community Containers $RELEASE"
 mkdir -p $DOWNLOAD_DIR && cd $DOWNLOAD_DIR
 
 echo "Downloading docker-compose file..."
-curl -f -O https://greenbone.github.io/docs/latest/_static/docker-compose-$RELEASE.yml
+curl -f -O https://greenbone.github.io/docs/latest/_static/docker-compose.yml
 
 echo "Pulling Greenbone Community Containers $RELEASE"
-docker compose -f $DOWNLOAD_DIR/docker-compose-$RELEASE.yml -p greenbone-community-edition pull
+docker compose -f $DOWNLOAD_DIR/docker-compose-$RELEASE.yml pull
 echo
 
 echo "Starting Greenbone Community Containers $RELEASE"
-docker compose -f $DOWNLOAD_DIR/docker-compose-$RELEASE.yml -p greenbone-community-edition up -d
+docker compose -f $DOWNLOAD_DIR/docker-compose-$RELEASE.yml up -d
 echo
 
 read -s -p "Password for admin user: " password
-docker compose -f $DOWNLOAD_DIR/docker-compose-$RELEASE.yml -p greenbone-community-edition \
+docker compose -f $DOWNLOAD_DIR/docker-compose-$RELEASE.yml \
     exec -u gvmd gvmd gvmd --user=admin --new-password=$password
 
 echo